This commit was generated by cvs2svn to compensate for changes in r13355, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@13356 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/petcat.c
@@ -984,7 +984,7 @@ static int p_expand(int version, int addr, int ctrls)
 
         /* prevent list protection from terminating listing */
 
-        while ((c = getc(source)) != EOF && !c);
+        while (((int)c = getc(source)) != EOF && !c);
 
         if (c == 0x12 && !line[2] && !line[3]) {  /* 00 00 12 22 */
             directory++;
@@ -1079,7 +1079,7 @@ static int p_expand(int version, int addr, int ctrls)
 
             _p_toascii((int)c, ctrls);  /* convert character */
 
-        } while ((c = getc(source)) != EOF && c);
+        } while (((int)c = getc(source)) != EOF && c);
         printf("\n");
 
     }      /* line */
